Move classes in sub packages
