+++ src/main/java/com/rettichlp/UnicacityAddon/commands/faction/badfaction/GaggedCommand.java
@@ -1,9 +1,9 @@
-package com.rettichlp.UnicacityAddon.commands;
+package com.rettichlp.UnicacityAddon.commands.faction.badfaction;
 
 import com.rettichlp.UnicacityAddon.base.abstraction.AbstractionLayer;
 import com.rettichlp.UnicacityAddon.base.abstraction.UPlayer;
 import com.rettichlp.UnicacityAddon.base.registry.annotation.UCCommand;
-import com.rettichlp.UnicacityAddon.base.utils.MathUtils;
+import com.rettichlp.UnicacityAddon.events.GaggedEventHandler;
 import net.minecraft.command.CommandBase;
 import net.minecraft.command.ICommandSender;
 import net.minecraft.server.MinecraftServer;
@@ -13,19 +13,18 @@ import java.util.Collections;
 import java.util.List;
 
 /**
- * @author RettichLP
+ * @author Dimiikou
  */
 @UCCommand
-public class AEquipCommand extends CommandBase {
+public class GaggedCommand extends CommandBase {
 
-    public static int amount = 0;
-
-    @Override @Nonnull public String getName() {
-        return "aequip";
+    @Override @Nonnull
+    public String getName() {
+        return "geknebelt";
     }
 
     @Override @Nonnull public String getUsage(@Nonnull ICommandSender sender) {
-        return "/aequip [Menge]";
+        return "/geknebelt";
     }
 
     @Override @Nonnull public List<String> getAliases() {
@@ -38,15 +37,9 @@ public class AEquipCommand extends CommandBase {
 
     @Override public void execute(@Nonnull MinecraftServer server, @Nonnull ICommandSender sender, @Nonnull String[] args) {
         UPlayer p = AbstractionLayer.getPlayer();
+        GaggedEventHandler.toggleGagged();
 
-        if (args.length > 0) {
-            if (!MathUtils.isInteger(args[0])) {
-                p.sendSyntaxMessage(getUsage(sender));
-                return;
-            }
-            amount = Integer.parseInt(args[0]);
-        }
-
-        p.sendInfoMessage("Menge für AEquip wurde eingestellt.");
+        if (GaggedEventHandler.isGagged()) p.sendInfoMessage("Ab sofort kannst du nur noch flüstern.");
+        else p.sendInfoMessage("Ab sofort kannst du wieder normal reden.");
     }
 }
